skills/pdf/requirements.txt
pypdf>=4.0.0
pdfplumber>=0.10.0
reportlab>=4.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
pytesseract>=0.3.10
pypdfium2>=4.0.0
pandas>=2.0.0
